refactor(tests): split chat-ipc.test.ts into harness + focused suites
3772-line chat-ipc.test.ts becomes a shared setupChatIpcTest()
harness (mock registry, 20 vi.mock registrations, fixtures) plus
three suites: core driver/gates (28 tests), lifecycle teardown/
cancel/queue (14), and compaction (10). Test names, bodies, and
assertions moved verbatim; 52/52 preserved, full suite 5279/5279.
Fixture dir is worker-scoped to avoid cross-file rm -rf races.
Docs/solutions regression-test pointers updated.
