+++ electron/tests/unit/chat-ipc.test.ts
@@ -1,822 +1,26 @@
-import * as fs from 'node:fs';
-import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
+import { setupChatIpcTest } from './chat-ipc-harness';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { IPC_CHANNELS } from '../../src/shared/types/ipc';
-import type { Agent } from '../../src/shared/types/agent';
 import { MessageRole, MessageType } from '../../src/shared/types/message';
-import { createCanonicalToolResult } from '../../src/shared/types/tool-result';
 import { ensureActiveSession } from '../../src/main/host/chat/session';
 import {
   discardDeletedSessionRuntime,
   forceAbortMainTurn,
 } from '../../src/main/host/chat/abort';
-import {
-  clearNextRequestStop,
-  requestNextRequestStop,
-  shouldStopNextRequest,
-} from '../../src/main/agents/next-request-stop';
 import { OPENAI_TIER_MECHANISM } from '../../src/main/providers/drivers/native';
 
-function successfulToolResult(toolCallId: string, content: string): Record<string, unknown> {
-  const canonical = createCanonicalToolResult('generic', {
-    status: 'complete',
-    data: { value: content },
-  });
-  return {
-    type: 'tool_result',
-    toolCallId,
-    content,
-    isError: false,
-    execution: {
-      canonical,
-      agentProjection: { content, completeness: 'complete' },
-    },
-  };
-}
-
-function cancelledToolResult(toolCallId: string): Record<string, unknown> {
-  const canonical = createCanonicalToolResult('generic', {
-    status: 'cancelled',
-    data: { questions: [], answers: [], cancelled: true },
-  });
-  return {
-    type: 'tool_result',
-    toolCallId,
-    content: 'Question cancelled',
-    isError: false,
-    execution: {
-      canonical,
-      agentProjection: { content: 'Question cancelled', completeness: 'complete' },
-    },
-  };
-}
-
-const mocks = vi.hoisted(() => {
-  const handlers = new Map<string, (...args: unknown[]) => unknown>();
-  const streamResponses: string[] = [];
-  const streamEventSequences: Array<Array<Record<string, unknown>>> = [];
-  let modelHistory: Array<Record<string, unknown>> = [];
-  // Real Electron resolves webContents.fromId(id) for the webContents serving
-  // an IPC request; emulate that registry so event delivery can address the
-  // inline sender objects by client id.
-  const sendersById = new Map<number, unknown>();
-  const electronWebContents = {
-    fromId: vi.fn((id: number) => sendersById.get(id) ?? null),
-    getAllWebContents: vi.fn(() => []),
-  };
-
-  type MockSession = {
-    id: string;
-    name: string;
-    model: string;
-    cwd: string | null;
-    chains: unknown[];
-    activeChainId: string | null;
-    createdAt: string;
-    updatedAt: string;
-    subagentChains: unknown[];
-    todoStore: { tasks: unknown[] };
-    selection?: { connectionId: string; modelId: string } | null;
-    modelLabel?: string;
-    reasoningEffortOverride?: string | number | null;
-    tierOverride?: string | null;
-  };
-  let activeSession: MockSession | null = null;
-  const sessionsById = new Map<string, MockSession>();
-  const activeSessionsByWindow = new Map<string, MockSession | null>();
-
-  let workspaceBound = true;
-  const testProjectDir = '/tmp/orchid-chat-ipc-project';
-  const workspaceByWindow = new Map<string, string>();
-  const generalAgent = {
-    name: 'general',
-    type: 'internal' as const,
-    tier: 'bloom' as const,
-    description: 'General-purpose agent',
-    system_prompt: 'You are a helpful assistant.',
-    allowed_tools: ['*'],
-    allowed_skills: ['*'],
-  } satisfies Agent;
-  const sessionNamerAgent = {
-    name: 'session-namer',
-    type: 'internal' as const,
-    tier: 'seed' as const,
-    description: 'Generates concise session titles',
-    system_prompt: 'Return one concise title for the supplied conversation.',
-    allowed_tools: [],
-    allowed_skills: [],
-  } satisfies Agent;
-  const runtimeByCwd = new Map<string, {
-    config: Record<string, unknown>;
-    agents: Map<string, Agent>;
-    skills: Map<string, unknown>;
-    personalities: Map<string, string>;
-    sharedPrompts: Record<'all-agents' | 'subagents', string | null>;
-  }>();
-  const runtimeRegistry = {
-    get: vi.fn((cwd: string) => {
-      const existing = runtimeByCwd.get(cwd);
-      if (existing) return { projectDir: cwd, ...existing };
-      return {
-        projectDir: cwd,
-        config: {
-          default_model: null,
-          tier_models: { bloom: null },
-          command_timeout: 30,
-          llm_stream_idle_timeout: 60,
-          llm_stream_retries: 0,
-          session_title_max_wait_seconds: 15,
-        },
-        agents: new Map([
-          ['general', generalAgent],
-          ['session-namer', sessionNamerAgent],
-        ]),
-        skills: new Map(),
-        personalities: new Map(),
-        sharedPrompts: { 'all-agents': null, subagents: null },
-      };
-    }),
-    _set: (cwd: string, runtime: {
-      config?: Record<string, unknown>;
-      agents?: Map<string, Agent>;
-      skills?: Map<string, unknown>;
-      personalities?: Map<string, string>;
-      sharedPrompts?: Record<'all-agents' | 'subagents', string | null>;
-    }) => {
-      runtimeByCwd.set(cwd, {
-        config: runtime.config ?? {
-          default_model: null,
-          tier_models: { bloom: null },
-          command_timeout: 30,
-          llm_stream_idle_timeout: 60,
-          llm_stream_retries: 0,
-          session_title_max_wait_seconds: 15,
-        },
-        agents: runtime.agents ?? new Map([
-          ['general', generalAgent],
-          ['session-namer', sessionNamerAgent],
-        ]),
-        skills: runtime.skills ?? new Map(),
-        personalities: runtime.personalities ?? new Map(),
-        sharedPrompts: runtime.sharedPrompts ?? { 'all-agents': null, subagents: null },
-      });
-    },
-    _reset: () => {
-      runtimeByCwd.clear();
-      runtimeRegistry.get.mockClear();
-    },
-  };
-  const toolRegistry = {
-    filter: vi.fn(() => []),
-    get: vi.fn(() => null),
-    validate: vi.fn(() => ({ ok: true as const, data: {} })),
-  };
-  const modelInstance = { provider: 'trusted-test-driver' };
-  const providerRuntime = {
-    resolveLanguageModel: vi.fn(async () => modelInstance),
-    resolveTierContext: vi.fn(async () => ({ connection: {}, tierMechanism: undefined })),
-    resolveExecution: vi.fn(async () => ({
-      modelInstance,
-      connection: {},
-      model: { id: 'vendor/path/model', capabilities: { reasoning: false } },
-      snapshot: {
-        providerId: 'openai',
-        providerDisplayName: 'OpenAI',
-        connectionId: '11111111-1111-4111-8111-111111111111',
-        connectionName: 'Work',
-        modelId: 'vendor/path/model',
-        protocol: 'openai-compatible',
-        modelSource: 'catalog',
-        catalogVersion: 1,
-        catalogSource: 'bundled',
-        catalogObservedAt: null,
-        pricing: null,
-        fieldProvenance: {},
-        statusObservation: null,
-      },
-    })),
-  };
-  const aiGenerateText = vi.fn(async () => ({ text: 'Investigate Session Naming' }));
-  // streamText adapter for the compaction paths: delegates to aiGenerateText so
-  // mockImplementationOnce handlers keep working, exposing the result through a
-  // single-chunk textStream + resolved usage (matches the consumption shape in
-  // summarize.ts / selective/run.ts).
-  const aiStreamText = vi.fn((args: Record<string, unknown>) => {
-    const ready = aiGenerateText(args) as Promise<{ text?: string }>;
-    // Zero-valued usage on both fulfilment and rejection so a failing
-    // aiGenerateText never surfaces as an unhandled rejection.
-    const zeroUsage = { inputTokens: 0, outputTokens: 0, totalTokens: 0 };
-    return {
-      textStream: (async function* () {
-        const result = await ready;
-        yield typeof result?.text === 'string' ? result.text : '';
-      })(),
-      usage: ready.then(() => zeroUsage, () => zeroUsage),
-    };
-  });
-  const wrappedTitleModel = { provider: 'wrapped-title-model' };
-  const aiWrapLanguageModel = vi.fn(() => wrappedTitleModel);
-  const createMiddlewareStack = vi.fn(() => []);
-  const accountingStore = {};
-  const buildSystemPromptContext = vi.fn(async ({ cwd }: { cwd: string }) => ({
-    cwd,
-    directoryTree: '',
-    subagents: [],
-    todos: [],
-    backgroundCommands: [],
-  }));
-  const mcpManager = {};
-
-  const sessionManager = {
-    getActive: vi.fn((windowId?: string) => (
-      windowId === undefined
-        ? activeSession
-        : activeSessionsByWindow.has(windowId)
-          ? activeSessionsByWindow.get(windowId) ?? null
-          : activeSession
-    )),
-    clearActive: vi.fn(() => {
-      activeSession = null;
-    }),
-    create: vi.fn((model: string | { connectionId: string; modelId: string }, options?: { cwd?: string | null }) => {
-      const modelLabel = typeof model === 'string' ? model : model.modelId;
-      activeSession = {
-        id: 'cccccccc-cccc-4ccc-8ccc-cccccccccccc',
-        name: 'Session draft',
-        model: modelLabel,
-        cwd: options?.cwd ?? null,
-        chains: [],
-        activeChainId: null,
-        createdAt: new Date().toISOString(),
-        updatedAt: new Date().toISOString(),
-        subagentChains: [],
-        todoStore: { tasks: [] },
-        selection: typeof model === 'string' ? undefined : model,
-        modelLabel,
-      };
-      sessionsById.set(activeSession.id, activeSession);
-      return activeSession;
-    }),
-    changeCwd: vi.fn((id: string, cwd: string) => {
-      if (!activeSession || activeSession.id !== id) {
-        throw new Error(`Cannot change cwd: session ${id} is not active`);
-      }
-      activeSession = { ...activeSession, cwd };
-      sessionsById.set(id, activeSession);
-      return activeSession;
-    }),
-    changeModel: vi.fn((
-      id: string,
-      selection: { connectionId: string; modelId: string },
-      modelLabel: string,
-    ) => {
-      if (!activeSession || activeSession.id !== id) {
-        throw new Error(`Cannot change model: session ${id} is not active`);
-      }
-      activeSession = { ...activeSession, selection, model: modelLabel, modelLabel };
-      sessionsById.set(id, activeSession);
-      return activeSession;
-    }),
-    setReasoningEffortOverride: vi.fn((id: string, effort: string | number | null) => {
-      const target = sessionsById.get(id) ?? (activeSession?.id === id ? activeSession : null);
-      if (!target) return;
-      const updated = { ...target, reasoningEffortOverride: effort };
-      sessionsById.set(id, updated);
-      if (activeSession?.id === id) activeSession = updated;
-    }),
-    setPermissionMode: vi.fn((id: string, mode: string | null) => {
-      const target = sessionsById.get(id) ?? (activeSession?.id === id ? activeSession : null);
-      if (!target) return;
-      const updated = { ...target, permissionMode: mode };
-      sessionsById.set(id, updated);
-      if (activeSession?.id === id) activeSession = updated;
-    }),
-    getSession: vi.fn((id: string) => sessionsById.get(id) ?? (activeSession?.id === id ? activeSession : null)),
-    load: vi.fn(() => null),
-    setCachedSession: vi.fn(),
-    /**
-     * Mirror of SessionManager.refreshCachedSessionFromStorage: the mock's
-     * "storage" is the sessionsById map, so the refresh returns the cached
-     * row (the mock does not model durable compaction splits — those are
-     * covered by session-compaction-persistence.test.ts).
-     */
-    refreshCachedSessionFromStorage: vi.fn((sessionId: string) =>
-      sessionsById.get(sessionId) ?? null),
-    getModelHistory: vi.fn(() => modelHistory),
-    switchTo: vi.fn((id: string) => {
-      const session = sessionsById.get(id) ?? (activeSession?.id === id ? activeSession : null);
-      if (session) activeSession = session;
-      return session;
-    }),
-    startChain: vi.fn((params?: { messages?: unknown[]; }, sessionId?: string) => {
-      const target = sessionId
-        ? sessionsById.get(sessionId) ?? null
-        : activeSession;
-      if (!target) return null;
-      const chain = {
-        id: `chain-${target.chains.length + 1}`,
-        sessionId: target.id,
-        messages: params?.messages ?? [],
-        status: 'active',
-        model: target.model,
-        agentName: 'general',
-        agentType: 'subagent',
-        agentTier: 'bloom',
-        subagentRecord: null,
-        startTime: new Date().toISOString(),
-        endTime: null,
-        errorDetail: null,
-        errorTitle: null,
-      };
-      const updated = {
-        ...target,
-        chains: [...target.chains, chain],
-        activeChainId: chain.id,
-      };
-      sessionsById.set(updated.id, updated);
-      if (activeSession?.id === updated.id) activeSession = updated;
-      return chain;
-    }),
-    updateActiveChainMessages: vi.fn((messages: unknown[], sessionId?: string) => {
-      const target = sessionId
-        ? sessionsById.get(sessionId) ?? null
-        : activeSession;
-      if (!target?.activeChainId) return null;
-      const updated = {
-        ...target,
-        chains: target.chains.map((chain: { id: string }) =>
-          chain.id === target.activeChainId ? { ...chain, messages } : chain
-        ),
-      };
-      sessionsById.set(updated.id, updated);
-      if (activeSession?.id === updated.id) activeSession = updated;
-      return updated;
-    }),
-    persistTurn: vi.fn((params: { messages: unknown[]; status?: string }, sessionId?: string) => {
-      const target = sessionId
-        ? sessionsById.get(sessionId) ?? null
-        : activeSession;
-      if (!target) return null;
-      const status = params.status ?? 'completed';
-      const activeId = target.activeChainId;
-      const idx = activeId
-        ? target.chains.findIndex((c: { id: string }) => c.id === activeId)
-        : -1;
-      let updated: MockSession;
-      if (idx >= 0) {
-        const chains = target.chains.map((c: { id: string }, i: number) =>
-          i === idx
-            ? {
-                ...c,
-                messages: params.messages,
-                status,
-                endTime: new Date().toISOString(),
-              }
-            : c,
-        );
-        updated = {
-          ...target,
-          chains,
-          activeChainId: null,
-        };
-      } else {
-        const chain = {
-          id: `chain-${target.chains.length + 1}`,
-          sessionId: target.id,
-          messages: params.messages,
-          status,
-          model: target.model,
-          agentName: 'general',
-          agentType: 'subagent',
-          agentTier: 'bloom',
-          subagentRecord: null,
-          startTime: new Date().toISOString(),
-          endTime: new Date().toISOString(),
-        };
-        updated = {
-          ...target,
-          chains: [...target.chains, chain],
-          activeChainId: null,
-        };
-      }
-      sessionsById.set(updated.id, updated);
-      if (activeSession?.id === updated.id) activeSession = updated;
-      return updated;
-    }),
-    /**
-     * Minimal mirror of SessionManager.applyCompaction (durable targeted
-     * compaction write): reports the post-write chain layout without touching
-     * the mock session rows — strict flag/anchor behavior is covered by
-     * session-compaction-persistence.test.ts against real storage.
-     */
-    applyCompaction: vi.fn((
-      sessionId: string,
-      payload: {
-        flaggedMessageIds?: readonly string[];
-        summaryChain?: { id: string } | null;
-      },
-    ) => {
-      const target = sessionsById.get(sessionId) ?? null;
-      if (!target) {
-        throw new Error(`applyCompaction: session ${sessionId} not found in durable rows`);
-      }
-      return {
-        chainIds: [
-          ...target.chains.map((chain: { id: string }) => chain.id),
-          ...(payload.summaryChain ? [payload.summaryChain.id] : []),
-        ],
-        flaggedChainIds: [],
-        summaryChainId: payload.summaryChain?.id ?? null,
-      };
-    }),
-    autoNameActive: vi.fn(async () => activeSession),
-    autoName: vi.fn(async (
-      _sessionId: string,
-      generateTitle?: (session: NonNullable<typeof activeSession>) => Promise<string | null>,
-    ) => {
-      if (!activeSession || !generateTitle) return activeSession;
-      const title = await generateTitle(activeSession);
-      if (title) {
-        activeSession = { ...activeSession, name: title };
-        sessionsById.set(activeSession.id, activeSession);
-      }
-      return activeSession;
-    }),
-    /** Test helper: reset between cases */
-    _reset: () => {
-      activeSession = null;
-      sessionsById.clear();
-      activeSessionsByWindow.clear();
-      modelHistory = [];
-      workspaceBound = true;
-      workspaceByWindow.clear();
-      sessionManager.getActive.mockClear();
-      sessionManager.create.mockClear();
-      sessionManager.changeCwd.mockClear();
-      sessionManager.changeModel.mockClear();
-      sessionManager.getSession.mockClear();
-      sessionManager.load.mockClear();
-      sessionManager.setCachedSession.mockClear();
-      sessionManager.getModelHistory.mockClear();
-      sessionManager.switchTo.mockClear();
-      sessionManager.clearActive.mockClear();
-      sessionManager.startChain.mockClear();
-      sessionManager.updateActiveChainMessages.mockClear();
-      sessionManager.persistTurn.mockClear();
-      sessionManager.applyCompaction.mockClear();
-      sessionManager.setReasoningEffortOverride.mockClear();
-      sessionManager.setPermissionMode.mockClear();
-      sessionManager.autoNameActive.mockClear();
-      sessionManager.autoName.mockClear();
-    },
-    /** Test helper: seed an active session without going through create(). */
-    _setActive: (session: MockSession | null) => {
-      activeSession = session;
-      if (session) sessionsById.set(session.id, session);
-    },
-    _setActiveForWindow: (windowId: string, session: MockSession | null) => {
-      activeSessionsByWindow.set(windowId, session);
-    },
-    /** Test helper: register a session without selecting it for the window. */
-    _putSession: (session: MockSession) => {
-      sessionsById.set(session.id, session);
-    },
-    _setModelHistory: (messages: Array<Record<string, unknown>>) => {
-      modelHistory = messages;
-    },
-  };
-
-  const workspace = {
-    resolveWorkspace: vi.fn((windowId: string, _options?: unknown) => {
-      if (!workspaceBound) {
-        return { cwd: null, source: 'unbound' as const, status: 'unbound' as const };
-      }
-      return {
-        cwd: workspaceByWindow.get(windowId) ?? testProjectDir,
-        source: 'default' as const,
-        status: 'valid' as const,
-      };
-    }),
-    isWorkspaceBound: vi.fn((info: { status: string; cwd: string | null }) => {
-      return info.status === 'valid' && info.cwd != null && info.cwd !== '';
-    }),
-    clearDraftCwd: vi.fn(),
-    setDraftCwd: vi.fn(),
-    getDraftCwd: vi.fn(() => null),
-    clearAllDraftCwds: vi.fn(),
-    updateStickyDefaultProjectDir: vi.fn(),
-    requireValidProjectDirectory: vi.fn((dir: string) => dir),
-    resolveWorkspaceFromParts: vi.fn(),
-    _setBound: (bound: boolean) => {
-      workspaceBound = bound;
-    },
-    _setWindowCwd: (windowId: string, cwd: string) => {
-      workspaceByWindow.set(windowId, cwd);
-    },
-    _testProjectDir: testProjectDir,
-  };
-
-  const backgroundEntries = new Map<number, {
-    sessionId: string | null;
-    agentScopeId: string;
-    tail: string;
-    exitCode: number | null;
-    interactive?: boolean;
-    owner?: 'AGENT' | 'USER';
-    command?: string;
-    description?: string;
-    createdAt?: number;
-    buffer?: { getTail: (lastN?: number) => string };
-  }>();
-  const backgroundStore = {
-    entries: backgroundEntries,
-    terminateSession: vi.fn(),
-    terminateScope: vi.fn(),
-    snapshot: vi.fn((commandId: number, _lastN?: number) => {
-      const entry = backgroundEntries.get(commandId);
-      if (!entry) return undefined;
-      return { tail: entry.tail, exitCode: entry.exitCode };
-    }),
-    snapshotVisible: vi.fn((
-      commandId: number,
-      _lastN?: number,
-      sessionId?: string | null,
-      agentScopeId?: string,
-    ) => {
-      const entry = backgroundEntries.get(commandId);
-      if (!entry) return undefined;
-      if (entry.sessionId !== (sessionId ?? null)) return undefined;
-      if (entry.agentScopeId !== (agentScopeId ?? 'main')) return undefined;
-      return { tail: entry.tail, exitCode: entry.exitCode };
-    }),
-    snapshotForSession: vi.fn((
-      commandId: number,
-      _lastN: number | undefined,
-      sessionId: string,
-    ) => {
-      const entry = backgroundEntries.get(commandId);
-      if (!entry || entry.sessionId !== sessionId) return undefined;
-      return { tail: entry.tail, exitCode: entry.exitCode };
-    }),
-    // The bgcmd:snapshot handler builds its found response directly from the
-    // entry (single-lookup path), so hydrate the fields minimal fixtures omit,
-    // mirroring real ProcessEntry defaults.
-    get: vi.fn((commandId: number) => {
-      const entry = backgroundEntries.get(commandId);
-      if (!entry) return undefined;
-      return {
-        interactive: false,
-        owner: 'AGENT' as const,
-        command: '',
-        ...entry,
-        buffer: entry.buffer ?? { getTail: () => entry.tail },
-      };
-    }),
-    list: vi.fn(() => []),
-    _reset: () => {
-      backgroundEntries.clear();
-      backgroundStore.terminateSession.mockClear();
-      backgroundStore.terminateScope.mockClear();
-      backgroundStore.snapshot.mockClear();
-      backgroundStore.snapshotVisible.mockClear();
-      backgroundStore.snapshotForSession.mockClear();
-      backgroundStore.get.mockClear();
-    },
-  };
-
-  return {
-    handlers,
-    streamResponses,
-    streamEventSequences,
-    sessionManager,
-    workspace,
-    takeDraftReasoningOverride: vi.fn(
-      () => undefined as string | number | null | undefined,
-    ),
-    takeDraftPermissionOverride: vi.fn(
-      () => undefined as string | null | undefined,
-    ),
-    ipcMain: {
-      handle: vi.fn((channel: string, handler: (...args: unknown[]) => unknown) => {
-        handlers.set(channel, (event: { sender?: { id?: number } }, ...rest: unknown[]) => {
-          const sender = event?.sender;
-          if (sender && typeof sender.id === 'number') sendersById.set(sender.id, sender);
-          return handler(event, ...rest);
-        });
-      }),
-      removeHandler: vi.fn((channel: string) => {
-        handlers.delete(channel);
-      }),
-    },
-    streamChat: vi.fn(async function* () {
-      const eventSequence = streamEventSequences.shift();
-      if (eventSequence) {
-        for (const event of eventSequence) {
-          yield event;
-        }
-        return;
-      }
-      const response = streamResponses.shift() ?? '';
-      if (response) {
-        yield { type: 'content', text: response };
-      }
-      yield { type: 'finish', finishReason: 'stop' };
-    }),
-    subagentManager: {
-      cancelRunning: vi.fn(() => []),
-      discardSession: vi.fn(),
-      getStates: vi.fn((): Array<{ id: string; state: string }> => []),
-      // U5: the embedded local host's HostServer subscribes to manager changes.
-      addOnChangeListener: vi.fn(() => vi.fn()),
-    },
-    sendersById,
-    publishSessionActivity: vi.fn(),
-    completeSessionActivity: vi.fn(),
-    electronWebContents,
-    runtimeRegistry,
-    generalAgent,
-    sessionNamerAgent,
-    toolRegistry,
-    modelInstance,
-    providerRuntime,
-    aiGenerateText,
-    aiStreamText,
-    aiWrapLanguageModel,
-    wrappedTitleModel,
-    createMiddlewareStack,
-    buildSystemPromptContext,
-    mcpManager,
-    accountingStore,
-    summarizeCompactableRange: vi.fn(),
-    saveSession: vi.fn(),
-    backgroundStore,
-  };
-});
-
-vi.mock('electron', () => ({
-  ipcMain: mocks.ipcMain,
-  BrowserWindow: {
-    getAllWindows: vi.fn(() => []),
-  },
-  webContents: mocks.electronWebContents,
-}));
-
-vi.mock('../../src/main/config/loader', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('../../src/main/config/loader')>();
-  return {
-    ...actual,
-    HOME_PERSONALITIES_DIR: '/tmp/orchid-test-personalities',
-    getTierModelSelection: (
-      config: {
-        default_model: unknown;
-        tier_models: Record<string, unknown>;
-      },
-      tier: string,
-    ) => config.tier_models[tier] ?? config.default_model,
-    getConfig: vi.fn(() => ({
-      default_model: null,
-      tier_models: { bloom: null },
-      command_timeout: 30,
-      llm_stream_idle_timeout: 60,
-      llm_stream_retries: 0,
-    })),
-  };
-});
-
-vi.mock('../../src/main/tools', () => ({
-  toolRegistry: mocks.toolRegistry,
-  createBuiltinToolRegistry: vi.fn(() => mocks.toolRegistry),
-  getBuiltinToolRegistryForRuntime: vi.fn(() => mocks.toolRegistry),
-  getSubagentManager: vi.fn(() => mocks.subagentManager),
-  // U5: the embedded local host's HostServer installs its own notifier.
-  setTodosChangedNotifier: vi.fn(),
-}));
-
-vi.mock('../../src/main/llm/orchestrator', () => ({
-  streamChat: mocks.streamChat,
-}));
-
-vi.mock('../../src/main/llm/compaction/summarize', async (importOriginal) => {
-  // Partial-module mock: keep every real export (resolveCompactorModelSelection
-  // uses the real fallback chain via the mocked config/loader) and override
-  // only the summarizer call itself.
-  const actual = await importOriginal<typeof import('../../src/main/llm/compaction/summarize')>();
-  return {
-    ...actual,
-    summarizeCompactableRange: mocks.summarizeCompactableRange,
-  };
-});
-
-vi.mock('../../src/main/session/storage', async (importOriginal) => {
-  // Partial-module mock: override only saveSession; every other storage
-  // export (loadSession, applyCompactionPersistence, …) stays real so
-  // session/manager dependencies resolve their actual implementations.
-  const actual = await importOriginal<typeof import('../../src/main/session/storage')>();
-  return {
-    ...actual,
-    saveSession: mocks.saveSession,
-  };
-});
-
-vi.mock('../../src/main/llm/middleware', () => ({
-  createMiddlewareStack: mocks.createMiddlewareStack,
-}));
-
-vi.mock('../../src/main/utils/esm-import', () => ({
-  importESM: vi.fn(async (specifier: string) => {
-    if (specifier === 'ai') {
-      return {
-        generateText: mocks.aiGenerateText,
-        streamText: mocks.aiStreamText,
-        wrapLanguageModel: mocks.aiWrapLanguageModel,
-      };
-    }
-    throw new Error(`Unexpected importESM specifier in chat IPC test: ${specifier}`);
-  }),
-}));
-
-vi.mock('../../src/main/providers', () => ({
-  getProviderRuntime: () => mocks.providerRuntime,
-}));
-
-vi.mock('../../src/main/providers/accounting/store', () => ({
-  getProviderAccountingStore: () => mocks.accountingStore,
-}));
-
-vi.mock('../../src/main/llm/build-prompt-context', () => ({
-  buildSystemPromptContext: mocks.buildSystemPromptContext,
-}));
-
-vi.mock('../../src/main/mcp/project-registry', () => ({
-  acquireProjectMCPManager: vi.fn(() => mocks.mcpManager),
-  releaseProjectMCPManager: vi.fn(),
-}));
-
-vi.mock('../../src/main/session/singleton', () => ({
-  getSessionManager: () => mocks.sessionManager,
-  resolveWindowWorkspace: (windowId: string) => mocks.workspace.resolveWorkspace(windowId),
-}));
-
-vi.mock('../../src/main/session/draft-reasoning', () => ({
-  takeDraftReasoningOverride: (windowId: string) =>
-    mocks.takeDraftReasoningOverride(windowId),
-}));
-
-vi.mock('../../src/main/permissions/session-overrides', () => ({
-  takeDraftPermissionOverride: (windowId: string) =>
-    mocks.takeDraftPermissionOverride(windowId),
-}));
-
-vi.mock('../../src/main/project/runtime', () => ({
-  getProjectRuntimeRegistry: () => mocks.runtimeRegistry,
-}));
-
-vi.mock('../../src/main/tools/process/background-store', () => ({
-  getBackgroundStore: () => mocks.backgroundStore,
-  setBackgroundStore: vi.fn(),
-  BackgroundProcessStore: vi.fn(),
-  subscribeBackgroundProcessChanges: vi.fn(() => vi.fn()),
-}));
-
-vi.mock('../../src/main/session/activity-live', () => ({
-  publishSessionActivity: mocks.publishSessionActivity,
-  completeSessionActivity: mocks.completeSessionActivity,
-  // U5: the embedded local host's HostServer installs its own broadcast.
-  setSessionActivityBroadcast: vi.fn(),
-}));
-
-// The trust gate is fail-closed for the mocked (non-existent) workspace dirs,
-// so these fixture cwds resolve as trusted to keep the suite on its own seams.
-vi.mock('../../src/main/project/trust', () => ({
-  getProjectTrustState: () => 'trusted',
-}));
-
-vi.mock('../../src/main/project/workspace', () => ({
-  resolveWorkspace: (...args: unknown[]) =>
-    mocks.workspace.resolveWorkspace(...(args as [string, unknown?])),
-  isWorkspaceBound: (...args: unknown[]) =>
-    mocks.workspace.isWorkspaceBound(...(args as [{ status: string; cwd: string | null }])),
-  clearDraftCwd: (...args: unknown[]) =>
-    mocks.workspace.clearDraftCwd(...(args as [string])),
-  setDraftCwd: mocks.workspace.setDraftCwd,
-  getDraftCwd: mocks.workspace.getDraftCwd,
-  clearAllDraftCwds: mocks.workspace.clearAllDraftCwds,
-  updateStickyDefaultProjectDir: mocks.workspace.updateStickyDefaultProjectDir,
-  requireValidProjectDirectory: mocks.workspace.requireValidProjectDirectory,
-  resolveWorkspaceFromParts: mocks.workspace.resolveWorkspaceFromParts,
-}));
+const {
+  mocks,
+  successfulToolResult,
+  cancelledToolResult,
+  doneEvents,
+  channelEvents,
+  waitForDoneCount,
+  waitForChannelCount,
+  makeSession,
+} = setupChatIpcTest();
 
 let chatIpc: typeof import('../../src/main/ipc/chat');
-
-// ensureActiveSession inspects the real fixture directory before the trust
-// gate (a deleted session folder must surface unbound_workspace), so the
-// fixture project path has to exist on disk.
-beforeAll(() => {
-  fs.mkdirSync(mocks.workspace._testProjectDir, { recursive: true });
-});
-
-afterAll(() => {
-  fs.rmSync(mocks.workspace._testProjectDir, { recursive: true, force: true });
-});
 
 describe('chat session selection gate', () => {
   beforeEach(async () => {
@@ -908,58 +112,6 @@ describe('chat session selection gate', () => {
     });
   });
 });
-
-function doneEvents(send: ReturnType<typeof vi.fn>) {
-  return send.mock.calls.filter(([channel]) => channel === IPC_CHANNELS.CHAT_DONE);
-}
-
-function channelEvents(send: ReturnType<typeof vi.fn>, channelName: string) {
-  return send.mock.calls.filter(([channel]) => channel === channelName);
-}
-
-async function waitForDoneCount(send: ReturnType<typeof vi.fn>, count: number) {
-  const deadline = Date.now() + 1000;
-
-  while (Date.now() < deadline) {
-    if (doneEvents(send).length >= count) {
-      return;
-    }
-    await new Promise((resolve) => setTimeout(resolve, 5));
-  }
-
-  throw new Error(`Timed out waiting for ${count} chat:done events`);
-}
-
-async function waitForChannelCount(
-  send: ReturnType<typeof vi.fn>,
-  channelName: string,
-  count: number,
-) {
-  const deadline = Date.now() + 1000;
-
-  while (Date.now() < deadline) {
-    if (channelEvents(send, channelName).length >= count) return;
-    await new Promise((resolve) => setTimeout(resolve, 5));
-  }
-
-  throw new Error(`Timed out waiting for ${count} ${channelName} event(s)`);
-}
-
-function makeSession(id: string, cwd = mocks.workspace._testProjectDir) {
-  return {
-    id,
-    name: `Session ${id.slice(0, 8)}`,
-    model: 'test/model',
-    cwd,
-    chains: [],
-    activeChainId: null,
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-    subagentChains: [],
-    todoStore: { tasks: [] },
-    permissionMode: null,
-  };
-}
 
 describe('chat IPC driver streaming', () => {
   beforeEach(async () => {
@@ -1494,7 +646,6 @@ describe('chat IPC driver streaming', () => {
     ]));
   });
 });
-
 
 describe('chat session snapshot hydration', () => {
   beforeEach(async () => {
@@ -2250,1523 +1401,5 @@ describe('chat IPC provider gates', () => {
     expect(mocks.aiGenerateText).not.toHaveBeenCalled();
     expect(channelEvents(send, IPC_CHANNELS.CHAT_DONE)).toEqual([]);
     expect(channelEvents(send, IPC_CHANNELS.CHAT_STATE)).toEqual([]);
-  });
-});
-
-describe('chat IPC teardown and bgcmd bounds', () => {
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    mocks.handlers.clear();
-    mocks.streamResponses.length = 0;
-    mocks.streamEventSequences.length = 0;
-    mocks.runtimeRegistry._reset();
-    mocks.sessionManager._reset();
-    chatIpc = await import('../../src/main/ipc/chat');
-    chatIpc.registerChatIPC();
-  });
-
-  afterEach(() => {
-    chatIpc.unregisterChatIPC();
-    mocks.handlers.clear();
-    mocks.sessionManager._reset();
-  });
-
-  it('unregisterChatIPC releases MCP project leases for active agents', async () => {
-    const projectRegistry = await import('../../src/main/mcp/project-registry');
-    const selection = {
-      connectionId: '11111111-1111-4111-8111-111111111111',
-      modelId: 'vendor/path/model',
-    };
-    mocks.sessionManager._setActive({
-      ...makeSession('cccccccc-cccc-4ccc-8ccc-cccccccccccc'),
-      selection,
-      modelLabel: selection.modelId,
-    });
-
-    let releaseStream: (() => void) | undefined;
-    const streamGate = new Promise<void>((resolve) => {
-      releaseStream = resolve;
-    });
-    mocks.streamChat.mockImplementationOnce(async function* () {
-      yield { type: 'content', text: 'partial' };
-      await streamGate;
-      yield { type: 'finish', finishReason: 'stop' };
-    });
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    const sendResult = await chatSend(
-      { sender: { id: 910, send } },
-      { message: 'Hold open for teardown' },
-    );
-    expect(sendResult).toMatchObject({ status: 'started' });
-
-    const deadline = Date.now() + 1000;
-    while (Date.now() < deadline && mocks.streamChat.mock.calls.length === 0) {
-      await new Promise((r) => setTimeout(r, 5));
-    }
-    expect(mocks.streamChat).toHaveBeenCalled();
-    expect(projectRegistry.acquireProjectMCPManager).toHaveBeenCalled();
-
-    chatIpc.unregisterChatIPC();
-    chatIpc.registerChatIPC();
-
-    expect(projectRegistry.releaseProjectMCPManager).toHaveBeenCalled();
-    releaseStream?.();
-  });
-
-  it('bgcmd:snapshot rejects lastN above the upper bound', async () => {
-    const snap = mocks.handlers.get(IPC_CHANNELS.BG_CMD_SNAPSHOT);
-    expect(snap).toBeDefined();
-
-    await expect(
-      snap!({ sender: { id: 911, send: vi.fn() } }, { commandId: 1, lastN: 1001 }),
-    ).rejects.toThrow(/Invalid bgcmd:snapshot/);
-  });
-
-  it('bgcmd:snapshot accepts lastN at the upper bound', async () => {
-    const snap = mocks.handlers.get(IPC_CHANNELS.BG_CMD_SNAPSHOT)!;
-
-    const result = await snap(
-      { sender: { id: 912, send: vi.fn() } },
-      { commandId: 999_999, lastN: 1000 },
-    );
-    expect(result).toEqual({ found: false });
-  });
-
-  it('bgcmd:snapshot denies cross-session command tails (M-P1-001)', async () => {
-    const ownerSession = 'aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa';
-    const otherSession = 'bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb';
-    mocks.backgroundStore.entries.set(42, {
-      sessionId: ownerSession,
-      agentScopeId: 'main',
-      tail: 'secret-output\n',
-      exitCode: 0,
-    });
-
-    mocks.sessionManager._setActive({
-      ...makeSession(otherSession),
-      selection: {
-        connectionId: '11111111-1111-4111-8111-111111111111',
-        modelId: 'vendor/path/model',
-      },
-    });
-
-    const snap = mocks.handlers.get(IPC_CHANNELS.BG_CMD_SNAPSHOT)!;
-    const denied = await snap(
-      { sender: { id: 913, send: vi.fn() } },
-      { commandId: 42, lastN: 50 },
-    );
-    expect(denied).toEqual({ found: false });
-
-    const allowed = await snap(
-      { sender: { id: 913, send: vi.fn() } },
-      { commandId: 42, lastN: 50, sessionId: ownerSession },
-    );
-    expect(allowed).toEqual({
-      found: true,
-      tail: 'secret-output\n',
-      exitCode: 0,
-      running: false,
-      interactive: false,
-      owner: 'AGENT',
-      command: '',
-      agentScopeId: 'main',
-    });
-  });
-
-  it('bgcmd:snapshot allows subagent-scoped tails within the same session', async () => {
-    const ownerSession = 'aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa';
-    mocks.backgroundStore.entries.set(43, {
-      sessionId: ownerSession,
-      agentScopeId: 'subagent-xyz',
-      tail: 'subagent-output\n',
-      exitCode: null,
-    });
-    mocks.sessionManager._setActive({
-      ...makeSession(ownerSession),
-      selection: {
-        connectionId: '11111111-1111-4111-8111-111111111111',
-        modelId: 'vendor/path/model',
-      },
-    });
-
-    const snap = mocks.handlers.get(IPC_CHANNELS.BG_CMD_SNAPSHOT)!;
-    const result = await snap(
-      { sender: { id: 914, send: vi.fn() } },
-      { commandId: 43, lastN: 50 },
-    );
-    expect(result).toEqual({
-      found: true,
-      tail: 'subagent-output\n',
-      exitCode: null,
-      running: true,
-      interactive: false,
-      owner: 'AGENT',
-      command: '',
-      agentScopeId: 'subagent-xyz',
-    });
-  });
-});
-
-describe('chat:cancel interrupt layers (issue #145)', () => {
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    mocks.handlers.clear();
-    mocks.streamResponses.length = 0;
-    mocks.streamEventSequences.length = 0;
-    mocks.subagentManager.cancelRunning.mockClear();
-    mocks.subagentManager.getStates.mockClear();
-    mocks.subagentManager.getStates.mockReturnValue([]);
-    mocks.runtimeRegistry._reset();
-    mocks.electronWebContents.fromId.mockReset();
-    mocks.sendersById.clear();
-    mocks.electronWebContents.fromId.mockImplementation(
-      (id: number) => mocks.sendersById.get(id) ?? null,
-    );
-    mocks.electronWebContents.getAllWebContents.mockReset();
-    mocks.electronWebContents.getAllWebContents.mockReturnValue([]);
-    mocks.sessionManager._reset();
-
-    chatIpc = await import('../../src/main/ipc/chat');
-    chatIpc.registerChatIPC();
-  });
-
-  afterEach(() => {
-    chatIpc.unregisterChatIPC();
-    mocks.handlers.clear();
-    mocks.streamResponses.length = 0;
-    mocks.streamEventSequences.length = 0;
-    mocks.sessionManager._reset();
-  });
-
-  it('second Esc terminates only main-scoped processes and leaves subagents running', async () => {
-    const sessionId = 'efefefef-efef-4efe-8efe-efefefefefef';
-    const selection = {
-      connectionId: '11111111-1111-4111-8111-111111111111',
-      modelId: 'vendor/path/model',
-    };
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      selection,
-      modelLabel: selection.modelId,
-    });
-    let releaseStream!: () => void;
-    const streamGate = new Promise<void>((resolve) => { releaseStream = resolve; });
-    mocks.streamChat.mockImplementationOnce(async function* () {
-      yield { type: 'content', text: 'subagents keep working' };
-      await streamGate;
-      yield { type: 'finish', finishReason: 'stop' };
-    });
-    const send = vi.fn();
-    const source = { id: 916, send };
-    mocks.sessionManager._setActiveForWindow('916', mocks.sessionManager.getActive()!);
-    mocks.electronWebContents.getAllWebContents.mockReturnValue([source]);
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    const chatCancel = mocks.handlers.get(IPC_CHANNELS.CHAT_CANCEL)!;
-
-    await chatSend({ sender: source }, { message: 'Run the crew' });
-    await waitForChannelCount(send, IPC_CHANNELS.CHAT_CHUNK, 1);
-
-    expect(await chatCancel({ sender: source }, { sessionId })).toMatchObject({ status: 'confirming' });
-    expect(await chatCancel({ sender: source }, { sessionId })).toMatchObject({ status: 'confirming_subagents' });
-
-    expect(mocks.backgroundStore.terminateScope).toHaveBeenCalledWith(sessionId, 'main');
-    expect(mocks.backgroundStore.terminateSession).not.toHaveBeenCalled();
-    expect(mocks.subagentManager.cancelRunning).not.toHaveBeenCalled();
-
-    releaseStream();
-    await new Promise((resolve) => setTimeout(resolve, 20));
-  });
-
-  it('keeps the subagent-cancel layer reachable after the main turn is gone', async () => {
-    const sessionId = 'fdfdfdfd-fdfd-4fdf-8fdf-fdfdfdfdfdfd';
-    const selection = {
-      connectionId: '11111111-1111-4111-8111-111111111111',
-      modelId: 'vendor/path/model',
-    };
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      selection,
-      modelLabel: selection.modelId,
-    });
-    mocks.streamResponses.push('turn finished');
-    const send = vi.fn();
-    const source = { id: 917, send };
-    mocks.sessionManager._setActiveForWindow('917', mocks.sessionManager.getActive()!);
-    mocks.electronWebContents.getAllWebContents.mockReturnValue([source]);
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    const chatCancel = mocks.handlers.get(IPC_CHANNELS.CHAT_CANCEL)!;
-
-    await chatSend({ sender: source }, { message: 'Delegate everything' });
-    await waitForChannelCount(send, IPC_CHANNELS.CHAT_DONE, 1);
-    // Let the finalize microtask dispose the ActiveAgent (post-interrupt-reset
-    // state: no live main turn, subagents still running).
-    await new Promise((resolve) => setTimeout(resolve, 20));
-
-    mocks.subagentManager.getStates.mockReturnValue([
-      { id: 'subagent-1', state: 'running' },
-    ]);
-    mocks.subagentManager.cancelRunning.mockReturnValue(['subagent-1']);
-    expect(await chatCancel({ sender: source }, { sessionId })).toMatchObject({ status: 'confirming_subagents' });
-    expect(mocks.subagentManager.cancelRunning).not.toHaveBeenCalled();
-
-    expect(await chatCancel({ sender: source }, { sessionId })).toMatchObject({ status: 'cancelled' });
-    expect(mocks.subagentManager.cancelRunning).toHaveBeenCalledWith(sessionId);
-    expect(mocks.backgroundStore.terminateSession).toHaveBeenCalledWith(sessionId);
-    expect(mocks.completeSessionActivity).toHaveBeenCalled();
-  });
-
-  it('returns no_active_stream when no main turn and no running subagents exist', async () => {
-    const sessionId = 'd1d1d1d1-d1d1-4d1d-8d1d-d1d1d1d1d1d1';
-    mocks.sessionManager._setActive(makeSession(sessionId));
-    const chatCancel = mocks.handlers.get(IPC_CHANNELS.CHAT_CANCEL)!;
-
-    expect(await chatCancel({ sender: { id: 918, send: vi.fn() } }, { sessionId }))
-      .toMatchObject({ status: 'no_active_stream' });
-    expect(mocks.subagentManager.cancelRunning).not.toHaveBeenCalled();
-  });
-});
-
-describe('chat:send draft single-flight (M-P1-013)', () => {
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    mocks.handlers.clear();
-    mocks.streamResponses.length = 0;
-    mocks.streamEventSequences.length = 0;
-    mocks.runtimeRegistry._reset();
-    mocks.sessionManager._reset();
-    chatIpc = await import('../../src/main/ipc/chat');
-    chatIpc.registerChatIPC();
-  });
-
-  afterEach(() => {
-    chatIpc.unregisterChatIPC();
-    mocks.handlers.clear();
-    mocks.sessionManager._reset();
-  });
-
-  it('concurrent draft sends create only one session', async () => {
-    const selection = {
-      connectionId: '11111111-1111-4111-8111-111111111111',
-      modelId: 'vendor/path/model',
-    };
-    // mockImplementationOnce (not mockImplementation): vi.clearAllMocks()
-    // clears calls but NOT implementations, so a persistent implementation here
-    // would shadow the harness default that serves later tests'
-    // streamResponses/streamEventSequences queues.
-    mocks.streamChat.mockImplementationOnce(async function* () {
-      yield { type: 'content', text: 'ok' };
-      await new Promise((r) => setTimeout(r, 30));
-      yield { type: 'finish', finishReason: 'stop' };
-    });
-
-    const send = vi.fn();
-    const webContents = { id: 920, send };
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-
-    const first = chatSend(
-      { sender: webContents },
-      { message: 'First draft send', model: selection },
-    );
-    const second = chatSend(
-      { sender: webContents },
-      { message: 'Second draft send', model: selection },
-    );
-    const [firstResult, secondResult] = await Promise.all([first, second]);
-
-    expect(mocks.sessionManager.create).toHaveBeenCalledTimes(1);
-    const statuses = [firstResult, secondResult].map(
-      (r) => (r as { status: string }).status,
-    );
-    expect(statuses.filter((s) => s === 'started')).toHaveLength(1);
-    expect(statuses).toContain('error');
-  });
-});
-
-describe('chat:queue_next early-stop signaling', () => {
-  const sessionId = 'ffffffff-ffff-4fff-8fff-ffffffffffff';
-
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    mocks.handlers.clear();
-    mocks.runtimeRegistry._reset();
-    mocks.sessionManager._reset();
-    clearNextRequestStop(sessionId);
-    chatIpc = await import('../../src/main/ipc/chat');
-    chatIpc.registerChatIPC();
-  });
-
-  afterEach(() => {
-    chatIpc.unregisterChatIPC();
-    mocks.handlers.clear();
-    clearNextRequestStop(sessionId);
-    mocks.sessionManager._reset();
-  });
-
-  it('sets the early-stop flag for a valid sessionId', async () => {
-    const queueNext = mocks.handlers.get(IPC_CHANNELS.CHAT_QUEUE_NEXT)!;
-    await queueNext({ sender: { id: 930, send: vi.fn() } }, { sessionId });
-    expect(shouldStopNextRequest(sessionId)).toBe(true);
-  });
-
-  it('is idempotent for repeated signals', async () => {
-    const queueNext = mocks.handlers.get(IPC_CHANNELS.CHAT_QUEUE_NEXT)!;
-    await queueNext({ sender: { id: 931, send: vi.fn() } }, { sessionId });
-    await queueNext({ sender: { id: 931, send: vi.fn() } }, { sessionId });
-    expect(shouldStopNextRequest(sessionId)).toBe(true);
-  });
-
-  it('throws on a payload missing sessionId', async () => {
-    const queueNext = mocks.handlers.get(IPC_CHANNELS.CHAT_QUEUE_NEXT)!;
-    await expect(
-      queueNext({ sender: { id: 932, send: vi.fn() } }, {}),
-    ).rejects.toThrow(/Invalid chat:queue_next payload/);
-    expect(shouldStopNextRequest(sessionId)).toBe(false);
-  });
-
-  it('throws on a wrong-typed sessionId', async () => {
-    const queueNext = mocks.handlers.get(IPC_CHANNELS.CHAT_QUEUE_NEXT)!;
-    await expect(
-      queueNext({ sender: { id: 933, send: vi.fn() } }, { sessionId: 123 }),
-    ).rejects.toThrow(/Invalid chat:queue_next payload/);
-    expect(shouldStopNextRequest(sessionId)).toBe(false);
-  });
-
-  it('clears the early-stop flag at chat:send turn start', async () => {
-    requestNextRequestStop(sessionId);
-    expect(shouldStopNextRequest(sessionId)).toBe(true);
-
-    const selection = {
-      connectionId: '11111111-1111-4111-8111-111111111111',
-      modelId: 'vendor/path/model',
-    };
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-    });
-    mocks.streamChat.mockImplementationOnce(async function* () {
-      yield { type: 'finish', finishReason: 'stop' };
-    });
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 934, send } }, { message: 'continue' });
-    await waitForDoneCount(send, 1);
-
-    expect(shouldStopNextRequest(sessionId)).toBe(false);
-  });
-});
-
-describe('chat compaction mid-turn pause', () => {
-  const selection = {
-    connectionId: '11111111-1111-4111-8111-111111111111',
-    modelId: 'vendor/path/model',
-  };
-
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    mocks.handlers.clear();
-    mocks.streamResponses.length = 0;
-    mocks.streamEventSequences.length = 0;
-    mocks.runtimeRegistry._reset();
-    mocks.sessionManager._reset();
-    mocks.runtimeRegistry._set(mocks.workspace._testProjectDir, {
-      config: {
-        default_model: null,
-        tier_models: { bloom: null },
-        command_timeout: 30,
-        llm_stream_idle_timeout: 60,
-        llm_stream_retries: 0,
-        session_title_max_wait_seconds: 0,
-        max_tool_steps: 100,
-        compaction: {
-          main: {
-            mode: 'simple',
-            threshold: 0.5,
-            model: null,
-            agent_name: 'compactor',
-            preserve_percent: 0.25,
-            min_compactable_tokens: 0,
-            mechanical_reclaim: true,
-            hysteresis_delta: 0.1,
-          },
-          subagents: {
-            mode: 'simple',
-            threshold: 0.85,
-            model: null,
-            agent_name: 'compactor-subagent',
-            preserve_percent: 0.25,
-            min_compactable_tokens: 4000,
-            mechanical_reclaim: true,
-            hysteresis_delta: 0.1,
-          },
-        },
-      },
-    });
-    mocks.providerRuntime.resolveExecution.mockImplementationOnce(async () => ({
-      modelInstance: mocks.modelInstance,
-      connection: {},
-      model: {
-        id: 'vendor/path/model',
-        capabilities: { reasoning: false },
-        limits: { contextTokens: 2000 },
-      },
-      snapshot: {
-        providerId: 'openai',
-        providerDisplayName: 'OpenAI',
-        connectionId: selection.connectionId,
-        connectionName: 'Work',
-        modelId: selection.modelId,
-        protocol: 'openai-compatible',
-        modelSource: 'catalog',
-        catalogVersion: 1,
-        catalogSource: 'bundled',
-        catalogObservedAt: null,
-        pricing: null,
-        fieldProvenance: {},
-        statusObservation: null,
-      },
-    }));
-    chatIpc = await import('../../src/main/ipc/chat');
-    chatIpc.registerChatIPC();
-  });
-
-  afterEach(() => {
-    chatIpc.unregisterChatIPC();
-    mocks.handlers.clear();
-    mocks.sessionManager._reset();
-  });
-
-  function compactionStream() {
-    return async function* () {
-      yield {
-        type: 'tool_call',
-        toolCallId: 'tc-compact-1',
-        toolName: 'read',
-        args: '{"path":"README.md"}',
-      };
-      yield successfulToolResult('tc-compact-1', 'x'.repeat(2000));
-      yield {
-        type: 'tool_call',
-        toolCallId: 'tc-compact-2',
-        toolName: 'read',
-        args: '{"path":"AGENTS.md"}',
-      };
-      yield successfulToolResult('tc-compact-2', 'y'.repeat(2000));
-      yield {
-        type: 'usage',
-        usage: {
-          prompt_tokens: 1800,
-          completion_tokens: 40,
-          total_tokens: 1840,
-          cached_tokens: 0,
-        },
-      };
-      yield { type: 'finish', finishReason: 'stop' };
-    };
-  }
-
-  it('resumes with accumulated turn history when the summary cannot be applied', async () => {
-    const sessionId = 'e0e0e0e0-e0e0-4e0e-8e0e-e0e0e0e0e0e0';
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-    });
-    mocks.summarizeCompactableRange.mockResolvedValueOnce(null);
-    mocks.streamChat.mockImplementationOnce(compactionStream());
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 940, send } }, { message: 'Explore with tools' });
-    await waitForDoneCount(send, 1);
-
-    expect(mocks.streamChat).toHaveBeenCalledTimes(2);
-    const resumedMessages = mocks.streamChat.mock.calls[1]![0]!.messages as Array<{
-      role: string;
-      tool_call_id?: string;
-    }>;
-    // The resumed replay must keep the turn's tool progress (the pre-fix code
-    // restarted from the bare turn base, collapsing context usage to the
-    // turn-start baseline with no compaction applied).
-    expect(resumedMessages.filter((m) => m.role === MessageRole.USER)).toHaveLength(1);
-    expect(resumedMessages.filter((m) => m.tool_call_id === 'tc-compact-1')).toHaveLength(2);
-    expect(resumedMessages.filter((m) => m.tool_call_id === 'tc-compact-2')).toHaveLength(2);
-    const compactionEvents = channelEvents(send, IPC_CHANNELS.CHAT_COMPACTION_PROGRESS)
-      .map(([, payload]) => payload);
-    expect(compactionEvents.length).toBeGreaterThanOrEqual(2);
-    // The window is resolvable, so the preparing widget event lands first
-    // (production always delivered it; the old fromId=null mock dropped it).
-    expect(compactionEvents[0]).toMatchObject({ phase: 'preparing', agentScopeId: 'main' });
-    expect(compactionEvents).toContainEqual(
-      expect.objectContaining({ phase: 'compacting', agentScopeId: 'main' }),
-    );
-    expect(compactionEvents.at(-1)).toMatchObject({ phase: 'complete', agentScopeId: 'main' });
-  });
-
-  it('applies the pending summary mid-turn and replays the compacted history', async () => {
-    const sessionId = 'e1e1e1e1-e1e1-4e1e-8e1e-e1e1e1e1e1e1';
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-    });
-    mocks.summarizeCompactableRange.mockResolvedValueOnce({
-      text: 'SUMMARY: explored compaction module',
-    });
-    mocks.streamChat.mockImplementationOnce(compactionStream());
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 941, send } }, { message: 'Explore with tools' });
-    await waitForDoneCount(send, 1);
-
-    expect(mocks.streamChat).toHaveBeenCalledTimes(2);
-    const resumedMessages = mocks.streamChat.mock.calls[1]![0]!.messages as Array<{
-      role: string;
-      content?: string;
-      compacted?: unknown;
-      excludeFromModel?: boolean;
-    }>;
-    const summary = resumedMessages.find((m) => m.compacted);
-    expect(summary?.content).toBe('SUMMARY: explored compaction module');
-    // Compacted range messages are excluded from replay but never deleted.
-    expect(resumedMessages.some((m) => m.tool_call_id === 'tc-compact-1' && m.excludeFromModel)).toBe(true);
-    // The durable targeted path is taken (never saveSession-from-view).
-    expect(mocks.sessionManager.applyCompaction).toHaveBeenCalledWith(
-      sessionId,
-      expect.objectContaining({
-        flaggedMessageIds: expect.arrayContaining([expect.any(String)]),
-        summaryChain: expect.objectContaining({ id: expect.any(String) }),
-      }),
-    );
-    expect(mocks.saveSession).not.toHaveBeenCalled();
-  });
-
-  it('persists the full turn (user message + pre-pause + post-resume content) into the durable active chain (P1 #3)', async () => {
-    const sessionId = 'e1e1e1e1-e1e1-4e1e-8e1e-e1e1e1e1e1f2';
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-    });
-    mocks.summarizeCompactableRange.mockResolvedValueOnce({
-      text: 'SUMMARY: explored compaction module',
-    });
-    // The resumed (second) stream call reads from streamResponses.
-    mocks.streamResponses.push('Resumed answer');
-    mocks.streamChat.mockImplementationOnce(compactionStream());
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 942, send } }, { message: 'Explore with tools' });
-    await waitForDoneCount(send, 1);
-
-    expect(mocks.streamChat).toHaveBeenCalledTimes(2);
-    const persisted = mocks.sessionManager.persistTurn.mock.calls.at(-1)?.[0] as {
-      messages: Array<Record<string, unknown>>;
-      status?: string;
-    };
-    expect(persisted.status).toBe('completed');
-    // The durable active chain keeps the FULL turn: the user message, the
-    // pre-pause tool progress, the inline summary head, and the post-resume
-    // assistant content (compaction inserts heads INLINE and only flags the
-    // prefix — one turn stays one chain row).
-    expect(persisted.messages).toEqual(expect.arrayContaining([
-      expect.objectContaining({ role: MessageRole.USER, content: 'Explore with tools' }),
-      expect.objectContaining({ type: MessageType.TOOL_CALL, tool_call_id: 'tc-compact-1' }),
-      expect.objectContaining({ role: MessageRole.ASSISTANT, content: 'Resumed answer' }),
-    ]));
-    // R31: simple-mode compaction never excludes user messages from the model
-    // replay (universal settle). The user message stays in the transcript,
-    // visible, and NOT flagged — the summary head stands in for the compacted
-    // non-user range, but the user intent survives verbatim.
-    const persistedUser = persisted.messages.find(
-      (m) => m.role === MessageRole.USER && m.content === 'Explore with tools',
-    )!;
-    expect(persistedUser.hidden).toBe(false);
-    expect(persistedUser.excludeFromModel).not.toBe(true);
-    expect(persisted.messages.some((m) => m.compacted)).toBe(true);
-    const done = doneEvents(send).at(-1)?.[1] as { messages: Array<Record<string, unknown>> };
-    expect(done.messages).toEqual(persisted.messages);
-  });
-
-  it('retries exactly once with compacted messages on a context-length error, then fails terminally (P1 #14)', async () => {
-    const sessionId = 'e6e6e6e6-e6e6-4e6e-8e6e-e6e6e6e6e6e6';
-    // R31: a user message alone is never compactable (pinned) — the retry
-    // path needs compactable assistant history to summarize.
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-      chains: [
-        {
-          id: 'chain-old',
-          messages: [
-            { id: 'a-old', role: 'assistant', content: 'y'.repeat(4000), type: 'text' },
-          ],
-        } as never,
-      ] as never,
-    });
-    mocks.sessionManager._setModelHistory([
-      { id: 'a-old', role: 'assistant', content: 'y'.repeat(4000), type: 'text' },
-    ]);
-    mocks.summarizeCompactableRange.mockResolvedValueOnce({ text: 'SUMMARY: retry compaction' });
-    // First attempt: provider reports a context-window overflow.
-    mocks.streamChat.mockImplementationOnce(async function* () {
-      yield {
-        type: 'error',
-        title: 'Stream Error',
-        detail: 'This model maximum context length is 2000 tokens; your input is too long',
-      };
-    });
-    // Retry attempt: overflow again — must terminate FAILED without a third call.
-    mocks.streamEventSequences.push([
-      {
-        type: 'error',
-        title: 'Stream Error',
-        detail: 'context_length_exceeded: input still too long after compaction',
-      },
-    ]);
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 943, send } }, { message: 'x'.repeat(4000) });
-    await waitForChannelCount(send, IPC_CHANNELS.CHAT_ERROR, 1);
-
-    // Exactly one retry (two streamChat calls) — never a third attempt.
-    expect(mocks.streamChat).toHaveBeenCalledTimes(2);
-    const retryMessages = mocks.streamChat.mock.calls[1]![0]!.messages as Array<{
-      content?: string;
-      compacted?: unknown;
-      excludeFromModel?: boolean;
-    }>;
-    // The retry request carries the compacted base: a summary head, the
-    // flagged assistant original (excluded from the model, never deleted), and
-    // the turn's user message — R31's universal settle keeps the user message
-    // in the model view, unflagged.
-    const summary = retryMessages.find((m) => m.compacted);
-    expect(summary?.content).toBe('SUMMARY: retry compaction');
-    expect(retryMessages.some((m) => m.content === 'y'.repeat(4000) && m.excludeFromModel)).toBe(true);
-    expect(retryMessages.some((m) => m.content === 'x'.repeat(4000) && !m.excludeFromModel)).toBe(true);
-
-    const failed = mocks.sessionManager.persistTurn.mock.calls.at(-1)?.[0] as {
-      messages: Array<Record<string, unknown>>;
-      status?: string;
-    };
-    expect(failed.status).toBe('failed');
-    // The failed turn anchors at the user message — R31 keeps it unflagged in
-    // the model view. The compacted base (flagged assistant original + summary
-    // head) lives in its own durable chains via the targeted applyCompaction
-    // write, never inside the turn slice.
-    expect(failed.messages.some((m) => m.role === MessageRole.USER && m.content === 'x'.repeat(4000) && !m.excludeFromModel)).toBe(true);
-    expect(failed.messages.some((m) => (m as { compacted?: unknown }).compacted)).toBe(false);
-    const error = channelEvents(send, IPC_CHANNELS.CHAT_ERROR).at(-1)?.[1] as {
-      kind?: string;
-      messages: Array<Record<string, unknown>>;
-    };
-    expect(error.kind).toBe('context_length_exceeded');
-    expect(error.messages).toEqual(failed.messages);
-    expect(mocks.sessionManager.applyCompaction).toHaveBeenCalledTimes(1);
-  });
-});
-
-describe('chat compaction selective pending (P1 #5)', () => {
-  const selection = {
-    connectionId: '11111111-1111-4111-8111-111111111111',
-    modelId: 'vendor/path/model',
-  };
-  // Substantive handoff texts (>= MIN_HANDOFF_SUMMARY_CHARS with real words) —
-  // the selective validator rejects degenerate activity-log texts for spans
-  // covering >= 1000 chars of source content.
-  const SELECTIVE_OLD_TURN_SUMMARY =
-    'SUMMARY: selective over old turn — the summarized assistant reply carried the findings, exact file paths, and decisions needed to continue; the next step flows into the preserved follow-up window without re-reading the old turn.';
-  const SELECTIVE_MID_TURN_SUMMARY =
-    'SUMMARY: selective mid-turn findings handoff — the summarized tool group preserved exact file paths, outcomes, and errors for the resumed stream; the next step continues the in-flight turn directly from the preserved window and open tool group.';
-  const compactorSelectiveAgent = {
-    name: 'compactor-selective',
-    type: 'internal' as const,
-    tier: 'seed' as const,
-    description: 'Selective compactor',
-    system_prompt: 'Return the selective ops JSON array.',
-    allowed_tools: [],
-    allowed_skills: [],
-  } satisfies Agent;
-
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    mocks.handlers.clear();
-    mocks.streamResponses.length = 0;
-    mocks.streamEventSequences.length = 0;
-    mocks.runtimeRegistry._reset();
-    mocks.sessionManager._reset();
-    // Clear any once-implementations leaked from earlier suites so the
-    // selective caller mock below is the only override in play.
-    mocks.aiGenerateText.mockReset();
-    mocks.aiGenerateText.mockImplementation(async () => ({ text: 'Investigate Session Naming' }));
-    mocks.runtimeRegistry._set(mocks.workspace._testProjectDir, {
-      config: {
-        default_model: null,
-        tier_models: { bloom: null },
-        command_timeout: 30,
-        llm_stream_idle_timeout: 60,
-        llm_stream_retries: 0,
-        session_title_max_wait_seconds: 0,
-        max_tool_steps: 100,
-        compaction: {
-          main: {
-            mode: 'selective',
-            threshold: 0.5,
-            model: null,
-            agent_name: 'compactor',
-            preserve_percent: 0.25,
-            min_compactable_tokens: 0,
-            mechanical_reclaim: true,
-            hysteresis_delta: 0.1,
-          },
-          subagents: {
-            mode: 'simple',
-            threshold: 0.85,
-            model: null,
-            agent_name: 'compactor-subagent',
-            preserve_percent: 0.25,
-            min_compactable_tokens: 4000,
-            mechanical_reclaim: true,
-            hysteresis_delta: 0.1,
-          },
-        },
-      },
-      agents: new Map<string, Agent>([
-        ['general', mocks.generalAgent],
-        ['session-namer', mocks.sessionNamerAgent],
-        ['compactor-selective', compactorSelectiveAgent],
-      ]),
-    });
-    mocks.providerRuntime.resolveExecution.mockImplementationOnce(async () => ({
-      modelInstance: mocks.modelInstance,
-      connection: {},
-      model: {
-        id: 'vendor/path/model',
-        capabilities: { reasoning: false },
-        limits: { contextTokens: 2000 },
-      },
-      snapshot: {
-        providerId: 'openai',
-        providerDisplayName: 'OpenAI',
-        connectionId: selection.connectionId,
-        connectionName: 'Work',
-        modelId: selection.modelId,
-        protocol: 'openai-compatible',
-        modelSource: 'catalog',
-        catalogVersion: 1,
-        catalogSource: 'bundled',
-        catalogObservedAt: null,
-        pricing: null,
-        fieldProvenance: {},
-        statusObservation: null,
-      },
-    }));
-    chatIpc = await import('../../src/main/ipc/chat');
-    chatIpc.registerChatIPC();
-  });
-
-  afterEach(() => {
-    chatIpc.unregisterChatIPC();
-    mocks.handlers.clear();
-    mocks.sessionManager._reset();
-  });
-
-  it('re-anchors the prepare-time selective replay so the next turn reaches the model intact', async () => {
-    const sessionId = 'e7e7e7e7-e7e7-4e7e-8e7e-e7e7e7e7e7e7';
-    // No `usage` on the fixture messages: a persisted usage would hydrate the
-    // trigger calibration and fire compaction SYNCHRONOUSLY at turn-1 send,
-    // which is the send-time path — not the pending this test exercises. The
-    // mid-turn usage event below is what prepares the selective pending.
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-      chains: [
-        {
-          id: 'chain-old',
-          messages: [
-            { id: 'u-old', role: 'user', content: 'x'.repeat(4000), type: 'text' },
-            { id: 'a-old', role: 'assistant', content: 'y'.repeat(4000), type: 'text' },
-          ],
-        } as never,
-      ] as never,
-    });
-    mocks.sessionManager._setModelHistory([
-      { id: 'u-old', role: 'user', content: 'x'.repeat(4000), type: 'text' },
-      { id: 'a-old', role: 'assistant', content: 'y'.repeat(4000), type: 'text' },
-    ]);
-    // The selective compactor keeps the user message verbatim and summarizes
-    // the assistant reply — ops derived from the manifest in the prompt. The
-    // summary text is a substantive handoff (the validator rejects activity
-    // logs for spans with >= 1000 chars of source).
-    mocks.aiGenerateText.mockImplementationOnce(async ({ messages }: { messages: Array<{ content: string }> }) => {
-      const prompt = String(messages[0]?.content ?? '');
-      const ids = [...prompt.matchAll(/^(\S+) \[/gm)].map((m) => m[1]!);
-      return {
-        text: JSON.stringify([
-          ...ids.filter((id) => id.startsWith('u-')).map((id) => ({ type: 'keep', id })),
-          { type: 'summarize', ids: ids.filter((id) => !id.startsWith('u-')), text: SELECTIVE_OLD_TURN_SUMMARY },
-        ]),
-      };
-    });
-    // Turn 1: the usage event prepares a selective pending; the provider error
-    // ends the turn before the pause can apply, so the pending survives to the
-    // next send (between-turns consumption).
-    mocks.streamChat.mockImplementationOnce(async function* () {
-      yield {
-        type: 'usage',
-        usage: { prompt_tokens: 1500, completion_tokens: 10, total_tokens: 1510, cached_tokens: 0 },
-      };
-      yield { type: 'error', title: 'Stream Error', detail: 'Provider disconnected' };
-    });
-    // Turn 2's stream reads from streamResponses.
-    mocks.streamResponses.push('Follow-up answer');
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 944, send } }, { message: 'First request' });
-    await waitForChannelCount(send, IPC_CHANNELS.CHAT_ERROR, 1);
-    await chatSend({ sender: { id: 944, send } }, { message: 'Follow up' });
-    await waitForDoneCount(send, 1);
-
-    // Turn 2's model request: the selective prefix (kept user message +
-    // synthetic summary) re-anchored onto the CURRENT history — the new user
-    // message and the previous turn's user message both reach the model. The
-    // pre-fix code spliced the prepare-time replay wholesale, so the new user
-    // message never reached the model and existingMessages stripped the
-    // previous user message instead.
-    const secondMessages = mocks.streamChat.mock.calls[1]![0]!.messages as Array<{
-      id?: string;
-      content?: string;
-      compacted?: unknown;
-    }>;
-    expect(secondMessages.some((m) => m.role === MessageRole.USER && m.content === 'Follow up')).toBe(true);
-    expect(secondMessages.some((m) => m.role === MessageRole.USER && m.content === 'First request')).toBe(true);
-    expect(secondMessages.some((m) => m.compacted && m.content === SELECTIVE_OLD_TURN_SUMMARY)).toBe(true);
-    expect(secondMessages.some((m) => m.id === 'u-old')).toBe(true);
-    // The summarized original is flagged durable-side and dropped from replay.
-    expect(secondMessages.some((m) => m.id === 'a-old')).toBe(false);
-
-    // Durable targeted write: the summarized original is flagged, the summary
-    // row inserted before the preserved window — never saveSession-from-view.
-    expect(mocks.sessionManager.applyCompaction).toHaveBeenCalledTimes(1);
-    expect(mocks.sessionManager.applyCompaction).toHaveBeenCalledWith(
-      sessionId,
-      expect.objectContaining({
-        flaggedMessageIds: ['a-old'],
-        // Unification #11b: synthesized compactor chains use one id scheme —
-        // randomUUID() — for both scopes (the main path previously synthesized
-        // `selective-${Date.now()}-…` ids here).
-        summaryChain: expect.objectContaining({
-          id: expect.stringMatching(/^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/),
-        }),
-        insertBeforeMessageId: expect.any(String),
-      }),
-    );
-    expect(mocks.saveSession).not.toHaveBeenCalled();
-
-    // The turn-2 durable active chain keeps the full turn.
-    const persisted = mocks.sessionManager.persistTurn.mock.calls.at(-1)?.[0] as {
-      messages: Array<Record<string, unknown>>;
-      status?: string;
-    };
-    expect(persisted.status).toBe('completed');
-    expect(persisted.messages).toEqual(expect.arrayContaining([
-      expect.objectContaining({ role: MessageRole.USER, content: 'Follow up' }),
-      expect.objectContaining({ role: MessageRole.ASSISTANT, content: 'Follow-up answer' }),
-    ]));
-  });
-
-  it('mid-turn selective apply keeps flagged originals + inline head in the durable turn row (review #54)', async () => {
-    const sessionId = 'e5e5e5e5-e5e5-4e5e-8e5e-e5e5e5e5e5e5';
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-    });
-    // Selective ops from the manifest: keep user + thinking verbatim,
-    // summarize every other span in one op.
-    mocks.aiGenerateText.mockImplementationOnce(async ({ messages }: { messages: Array<{ content: string }> }) => {
-      const prompt = String(messages[0]?.content ?? '');
-      const entries = [...prompt.matchAll(/^(\S+) \[([\w-]+)\]/gm)].map((m) => ({
-        id: m[1]!,
-        kind: m[2]!,
-      }));
-      return {
-        text: JSON.stringify([
-          ...entries.filter((e) => e.kind === 'user' || e.kind === 'thinking')
-            .map((e) => ({ type: 'keep', id: e.id })),
-          {
-            type: 'summarize',
-            ids: entries.filter((e) => e.kind !== 'user' && e.kind !== 'thinking').map((e) => e.id),
-            text: SELECTIVE_MID_TURN_SUMMARY,
-          },
-        ]),
-      };
-    });
-    // First stream: two tool groups + usage over threshold → selective pending
-    // prepares; the turn pauses, applies, and resumes. The resume generator
-    // below HOLDS until the 300ms checkpoint debounce scheduled by the
-    // triggering usage event has actually fired post-apply — exactly the
-    // stale-capture race window (review #55).
-    mocks.streamChat.mockImplementationOnce(async function* () {
-      yield {
-        type: 'tool_call',
-        toolCallId: 'tc-sel-1',
-        toolName: 'read',
-        args: '{"path":"README.md"}',
-      };
-      yield successfulToolResult('tc-sel-1', 'x'.repeat(2000));
-      yield {
-        type: 'tool_call',
-        toolCallId: 'tc-sel-2',
-        toolName: 'read',
-        args: '{"path":"AGENTS.md"}',
-      };
-      yield successfulToolResult('tc-sel-2', 'y'.repeat(2000));
-      yield {
-        type: 'usage',
-        usage: { prompt_tokens: 1800, completion_tokens: 40, total_tokens: 1840, cached_tokens: 0 },
-      };
-      yield { type: 'finish', finishReason: 'stop' };
-    });
-    mocks.streamChat.mockImplementationOnce(async function* () {
-      // Deterministic handshake instead of a fixed sleep: wait until a
-      // checkpoint write (updateActiveChainMessages) has fired AFTER the
-      // compaction apply — the debounced checkpoint landing mid-resume —
-      // bounded by an explicit timeout budget.
-      const applyOrders = mocks.sessionManager.applyCompaction.mock.invocationCallOrder;
-      const updateOrders = mocks.sessionManager.updateActiveChainMessages.mock.invocationCallOrder;
-      const checkpointFiredPostApply = () => {
-        const applyOrder = applyOrders[0];
-        return applyOrder !== undefined && updateOrders.some((order) => order > applyOrder);
-      };
-      const deadline = Date.now() + 1000;
-      while (!checkpointFiredPostApply() && Date.now() < deadline) {
-        await new Promise((resolve) => setTimeout(resolve, 5));
-      }
-      yield { type: 'content', text: 'Resumed answer' };
-      yield { type: 'finish', finishReason: 'stop' };
-    });
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 945, send } }, { message: 'Explore with tools' });
-    await waitForDoneCount(send, 1);
-
-    expect(mocks.streamChat).toHaveBeenCalledTimes(2);
-    expect(mocks.sessionManager.applyCompaction).toHaveBeenCalledTimes(1);
-
-    // The resumed model view is the reanchored REPLAY: the head reaches the
-    // model, the flagged original does not.
-    const resumedMessages = mocks.streamChat.mock.calls[1]![0]!.messages as Array<{
-      content?: string;
-      compacted?: unknown;
-      excludeFromModel?: boolean;
-      tool_call_id?: string;
-    }>;
-    expect(resumedMessages.some((m) => m.compacted && m.content === SELECTIVE_MID_TURN_SUMMARY)).toBe(true);
-    expect(resumedMessages.some((m) => m.tool_call_id === 'tc-sel-1' && m.excludeFromModel)).toBe(false);
-
-    // The durable turn row (finalize rewrite input) is TRANSCRIPT-complete:
-    // user message unflagged, the flagged ORIGINAL tool pair preserved with
-    // excludeFromModel, the inline summary head, and the post-resume content.
-    // The pre-fix code rewrote the row from the model-view replay and
-    // durably erased the flagged originals + head (review #54).
-    const persisted = mocks.sessionManager.persistTurn.mock.calls.at(-1)?.[0] as {
-      messages: Array<Record<string, unknown>>;
-      status?: string;
-    };
-    expect(persisted.status).toBe('completed');
-    const persistedUser = persisted.messages.find(
-      (m) => m.role === MessageRole.USER && m.content === 'Explore with tools',
-    )!;
-    expect(persistedUser.excludeFromModel).not.toBe(true);
-    expect(persisted.messages.some((m) => m.tool_call_id === 'tc-sel-1' && m.excludeFromModel === true)).toBe(true);
-    expect(persisted.messages.some((m) => m.compacted && (m as { content?: string }).content === SELECTIVE_MID_TURN_SUMMARY)).toBe(true);
-    expect(persisted.messages.some((m) => m.role === MessageRole.ASSISTANT && m.content === 'Resumed answer')).toBe(true);
-
-    // Stale-checkpoint race (review #55): a checkpoint scheduled by the same
-    // usage event that triggered the compaction fires after the apply — its
-    // write must reflect the compacted transcript (re-derived at fire time),
-    // never the pre-compaction snapshot captured at schedule time.
-    const applyOrder = mocks.sessionManager.applyCompaction.mock.invocationCallOrder[0]!;
-    const updateMock = mocks.sessionManager.updateActiveChainMessages;
-    const postApplyIndexes = updateMock.mock.invocationCallOrder
-      .map((order, index) => (order > applyOrder ? index : -1))
-      .filter((index) => index >= 0);
-    // The race is only exercised when a checkpoint write actually fired after
-    // the apply — fail loudly instead of passing vacuously when none did.
-    expect(postApplyIndexes.length).toBeGreaterThan(0);
-    for (const index of postApplyIndexes) {
-      const written = updateMock.mock.calls[index]![0] as Array<{
-        tool_call_id?: string;
-        excludeFromModel?: boolean;
-      }>;
-      expect(written.some((m) => m.tool_call_id === 'tc-sel-1' && m.excludeFromModel === true)).toBe(true);
-    }
-  });
-});
-
-describe('chat compaction disabled without a model context limit', () => {
-  const selection = {
-    connectionId: '11111111-1111-4111-8111-111111111111',
-    modelId: 'vendor/path/model',
-  };
-
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    mocks.handlers.clear();
-    mocks.streamResponses.length = 0;
-    mocks.streamEventSequences.length = 0;
-    mocks.runtimeRegistry._reset();
-    mocks.sessionManager._reset();
-    mocks.runtimeRegistry._set(mocks.workspace._testProjectDir, {
-      config: {
-        default_model: null,
-        tier_models: { bloom: null },
-        command_timeout: 30,
-        llm_stream_idle_timeout: 60,
-        llm_stream_retries: 0,
-        session_title_max_wait_seconds: 0,
-        max_tool_steps: 100,
-        compaction: {
-          main: {
-            mode: 'simple',
-            threshold: 0.5,
-            model: null,
-            agent_name: 'compactor',
-            preserve_percent: 0.25,
-            min_compactable_tokens: 0,
-            mechanical_reclaim: true,
-            hysteresis_delta: 0.1,
-          },
-          subagents: {
-            mode: 'simple',
-            threshold: 0.85,
-            model: null,
-            agent_name: 'compactor-subagent',
-            preserve_percent: 0.25,
-            min_compactable_tokens: 4000,
-            mechanical_reclaim: true,
-            hysteresis_delta: 0.1,
-          },
-        },
-      },
-    });
-    // Default resolveExecution mock returns no model.limits → contextTokens null.
-    chatIpc = await import('../../src/main/ipc/chat');
-    chatIpc.registerChatIPC();
-  });
-
-  afterEach(() => {
-    chatIpc.unregisterChatIPC();
-    mocks.handlers.clear();
-    mocks.sessionManager._reset();
-  });
-
-  function hugeHistoryStream() {
-    return async function* () {
-      yield {
-        type: 'tool_call',
-        toolCallId: 'tc-nolimit-1',
-        toolName: 'read',
-        args: '{"path":"README.md"}',
-      };
-      yield successfulToolResult('tc-nolimit-1', 'x'.repeat(500_000));
-      yield {
-        type: 'usage',
-        usage: {
-          prompt_tokens: 120_000,
-          completion_tokens: 40,
-          total_tokens: 120_040,
-          cached_tokens: 0,
-        },
-      };
-      yield { type: 'finish', finishReason: 'stop' };
-    };
-  }
-
-  it('never compacts on send when the model has no configured context window', async () => {
-    const sessionId = 'e2e2e2e2-e2e2-4e2e-8e2e-e2e2e2e2e2e2';
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-    });
-    mocks.streamChat.mockImplementationOnce(hugeHistoryStream());
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 950, send } }, { message: 'Explore with tools' });
-    await waitForDoneCount(send, 1);
-
-    await chatSend({ sender: { id: 950, send } }, { message: 'Follow up' });
-    await waitForDoneCount(send, 2);
-
-    // The pre-fix send-time path substituted an assumed 128k window for the
-    // null limit and compacted this history; compaction must stay disabled.
-    expect(mocks.summarizeCompactableRange).not.toHaveBeenCalled();
-    expect(mocks.streamChat).toHaveBeenCalledTimes(2);
-    const secondMessages = mocks.streamChat.mock.calls[1]![0]!.messages as Array<{
-      content?: string;
-      compacted?: unknown;
-      excludeFromModel?: boolean;
-    }>;
-    expect(secondMessages.some((m) => m.compacted)).toBe(false);
-    expect(secondMessages.some((m) => m.excludeFromModel)).toBe(false);
-    // The oversized tool result is replayed verbatim — no reclaim offload.
-    expect(secondMessages.some((m) => typeof m.content === 'string' && m.content.length > 400_000)).toBe(true);
-  });
-});
-
-describe('chat compaction send-time calibration', () => {
-  const selection = {
-    connectionId: '11111111-1111-4111-8111-111111111111',
-    modelId: 'vendor/path/model',
-  };
-
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    mocks.handlers.clear();
-    mocks.streamResponses.length = 0;
-    mocks.streamEventSequences.length = 0;
-    mocks.runtimeRegistry._reset();
-    mocks.sessionManager._reset();
-    mocks.runtimeRegistry._set(mocks.workspace._testProjectDir, {
-      config: {
-        default_model: null,
-        tier_models: { bloom: null },
-        command_timeout: 30,
-        llm_stream_idle_timeout: 60,
-        llm_stream_retries: 0,
-        session_title_max_wait_seconds: 0,
-        max_tool_steps: 100,
-        compaction: {
-          main: {
-            mode: 'simple',
-            threshold: 0.5,
-            model: null,
-            agent_name: 'compactor',
-            preserve_percent: 0.25,
-            min_compactable_tokens: 0,
-            mechanical_reclaim: true,
-            hysteresis_delta: 0.1,
-          },
-          subagents: {
-            mode: 'simple',
-            threshold: 0.85,
-            model: null,
-            agent_name: 'compactor-subagent',
-            preserve_percent: 0.25,
-            min_compactable_tokens: 4000,
-            mechanical_reclaim: true,
-            hysteresis_delta: 0.1,
-          },
-        },
-      },
-    });
-    mocks.providerRuntime.resolveExecution.mockImplementationOnce(async () => ({
-      modelInstance: mocks.modelInstance,
-      connection: {},
-      model: {
-        id: 'vendor/path/model',
-        capabilities: { reasoning: false },
-        limits: { contextTokens: 2000 },
-      },
-      snapshot: {
-        providerId: 'openai',
-        providerDisplayName: 'OpenAI',
-        connectionId: selection.connectionId,
-        connectionName: 'Work',
-        modelId: selection.modelId,
-        protocol: 'openai-compatible',
-        modelSource: 'catalog',
-        catalogVersion: 1,
-        catalogSource: 'bundled',
-        catalogObservedAt: null,
-        pricing: null,
-        fieldProvenance: {},
-        statusObservation: null,
-      },
-    }));
-    chatIpc = await import('../../src/main/ipc/chat');
-    chatIpc.registerChatIPC();
-  });
-
-  afterEach(() => {
-    chatIpc.unregisterChatIPC();
-    mocks.handlers.clear();
-    mocks.sessionManager._reset();
-  });
-
-  function textOnlyStream() {
-    return async function* () {
-      yield { type: 'text', data: 'ok' };
-      yield { type: 'finish', finishReason: 'stop' };
-    };
-  }
-
-  it('skips send-time compaction when uncalibrated, even over threshold (hard rule: no chars/4)', async () => {
-    const sessionId = 'e3e3e3e3-e3e3-4e3e-8e3e-e3e3e3e3e3e3';
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-    });
-    // Large prior history, but NO usage observation anywhere: no context
-    // snapshots (store unavailable in tests) and no chain message usages.
-    mocks.sessionManager._setModelHistory([
-      { id: 'u-old', role: 'user', content: 'x'.repeat(4000), type: 'text' },
-    ]);
-    mocks.streamChat.mockImplementationOnce(textOnlyStream());
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 960, send } }, { message: 'Follow up' });
-    await waitForDoneCount(send, 1);
-
-    // chars/4 would estimate ~1000 tokens = threshold(0.5) × 2000 and fire;
-    // the hard rule keeps the estimate unknown and skips proactive compaction.
-    expect(mocks.summarizeCompactableRange).not.toHaveBeenCalled();
-    expect(mocks.streamChat).toHaveBeenCalledTimes(1);
-    const sentMessages = mocks.streamChat.mock.calls[0]![0]!.messages as Array<{
-      content?: string;
-      compacted?: unknown;
-    }>;
-    expect(sentMessages.some((m) => m.compacted)).toBe(false);
-    expect(sentMessages.some((m) => m.content === 'x'.repeat(4000))).toBe(true);
-  });
-
-  it('hydrates calibration from persisted chain usage and compacts within the preserve budget', async () => {
-    const sessionId = 'e4e4e4e4-e4e4-4e4e-8e4e-e4e4e4e4e4e4';
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-      chains: [
-        {
-          id: 'chain-old',
-          messages: [
-            // R31: the pinned user message survives every compaction verbatim;
-            // the oversized assistant reply is the compactable range.
-            { id: 'u-old', role: 'user', content: 'x'.repeat(4000), type: 'text', usage: { prompt_tokens: 1500 } },
-            { id: 'a-old', role: 'assistant', content: 'y'.repeat(4000), type: 'text' },
-          ],
-        },
-      ] as never,
-    });
-    mocks.sessionManager._setModelHistory([
-      { id: 'u-old', role: 'user', content: 'x'.repeat(4000), type: 'text' },
-      { id: 'a-old', role: 'assistant', content: 'y'.repeat(4000), type: 'text' },
-    ]);
-    mocks.summarizeCompactableRange.mockResolvedValueOnce({ text: 'SUMMARY: prior turn' });
-    mocks.streamChat.mockImplementationOnce(textOnlyStream());
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 961, send } }, { message: 'Follow up' });
-    await waitForDoneCount(send, 1);
-
-    // Hydrated lastObserved=1500 → ratio ≈ 0.37 → estimate 1500 = 0.75 of the
-    // 2000 window ≥ 0.5 threshold → fires with preserve budget 500 tokens.
-    expect(mocks.summarizeCompactableRange).toHaveBeenCalledTimes(1);
-    expect(mocks.streamChat).toHaveBeenCalledTimes(1);
-    const sentMessages = mocks.streamChat.mock.calls[0]![0]!.messages as Array<{
-      content?: string;
-      compacted?: unknown;
-      excludeFromModel?: boolean;
-    }>;
-    const summary = sentMessages.find((m) => m.compacted);
-    expect(summary?.content).toBe('SUMMARY: prior turn');
-    // R31: the oversized prior message is a user message, so it stays in the
-    // model view verbatim — never flagged, never deleted. The summary head
-    // stands in for the compacted range, not for the user's intent.
-    const oldMessage = sentMessages.find((m) => m.content === 'x'.repeat(4000));
-    expect(oldMessage).toBeDefined();
-    expect(oldMessage?.excludeFromModel).not.toBe(true);
-    // The summarized assistant original is excluded from replay, never deleted.
-    expect(sentMessages.some((m) => m.content === 'y'.repeat(4000) && m.excludeFromModel)).toBe(true);
-    // The durable targeted path is taken (never saveSession-from-view). R31's
-    // universal settle filters user ids from the flagged set.
-    expect(mocks.sessionManager.applyCompaction).toHaveBeenCalledWith(
-      sessionId,
-      expect.objectContaining({
-        flaggedMessageIds: ['a-old'],
-        summaryChain: expect.objectContaining({ id: expect.any(String) }),
-      }),
-    );
-    expect(mocks.saveSession).not.toHaveBeenCalled();
-  });
-});
-
-describe('chat compaction scoped user settle (keep_last_user_messages)', () => {
-  const selection = {
-    connectionId: '11111111-1111-4111-8111-111111111111',
-    modelId: 'vendor/path/model',
-  };
-
-  beforeEach(async () => {
-    vi.clearAllMocks();
-    mocks.handlers.clear();
-    mocks.streamResponses.length = 0;
-    mocks.streamEventSequences.length = 0;
-    mocks.runtimeRegistry._reset();
-    mocks.sessionManager._reset();
-    mocks.runtimeRegistry._set(mocks.workspace._testProjectDir, {
-      config: {
-        default_model: null,
-        tier_models: { bloom: null },
-        command_timeout: 30,
-        llm_stream_idle_timeout: 60,
-        llm_stream_retries: 0,
-        session_title_max_wait_seconds: 0,
-        max_tool_steps: 100,
-        compaction: {
-          main: {
-            mode: 'simple',
-            threshold: 0.5,
-            model: null,
-            agent_name: 'compactor',
-            preserve_percent: 0.25,
-            min_compactable_tokens: 0,
-            mechanical_reclaim: true,
-            hysteresis_delta: 0.1,
-            keep_last_user_messages: 10,
-            pin_first_user_message: false,
-          },
-          subagents: {
-            mode: 'simple',
-            threshold: 0.85,
-            model: null,
-            agent_name: 'compactor-subagent',
-            preserve_percent: 0.25,
-            min_compactable_tokens: 4000,
-            mechanical_reclaim: true,
-            hysteresis_delta: 0.1,
-          },
-        },
-      },
-    });
-    mocks.providerRuntime.resolveExecution.mockImplementationOnce(async () => ({
-      modelInstance: mocks.modelInstance,
-      connection: {},
-      model: {
-        id: 'vendor/path/model',
-        capabilities: { reasoning: false },
-        limits: { contextTokens: 2000 },
-      },
-      snapshot: {
-        providerId: 'openai',
-        providerDisplayName: 'OpenAI',
-        connectionId: selection.connectionId,
-        connectionName: 'Work',
-        modelId: selection.modelId,
-        protocol: 'openai-compatible',
-        modelSource: 'catalog',
-        catalogVersion: 1,
-        catalogSource: 'bundled',
-        catalogObservedAt: null,
-        pricing: null,
-        fieldProvenance: {},
-        statusObservation: null,
-      },
-    }));
-    chatIpc = await import('../../src/main/ipc/chat');
-    chatIpc.registerChatIPC();
-  });
-
-  afterEach(() => {
-    chatIpc.unregisterChatIPC();
-    mocks.handlers.clear();
-    mocks.sessionManager._reset();
-  });
-
-  function textOnlyStream() {
-    return async function* () {
-      yield { type: 'text', data: 'ok' };
-      yield { type: 'finish', finishReason: 'stop' };
-    };
-  }
-
-  it('flags the 11th-oldest user message (keep_last=10): out of the model view, carried only by the summary', async () => {
-    const sessionId = 'e5e5e5e5-e5e5-4e5e-8e5e-e5e5e5e5e5e5';
-    // Ten user/assistant fixture turns; the sent follow-up makes eleven user
-    // messages. keep_last=10 + pin_first=false → only the ten newest users
-    // are exempt, so u-0 can leave the model view for the first time.
-    const fixtureMessages: Array<Record<string, unknown>> = [];
-    for (let i = 0; i < 10; i += 1) {
-      fixtureMessages.push({ id: `u-${i}`, role: 'user', content: 'u'.repeat(200), type: 'text' });
-      fixtureMessages.push({ id: `a-${i}`, role: 'assistant', content: 'a'.repeat(400), type: 'text' });
-    }
-    fixtureMessages.at(-1)!.usage = { prompt_tokens: 3000 };
-    mocks.sessionManager._setActive({
-      ...makeSession(sessionId),
-      model: selection.modelId,
-      selection,
-      modelLabel: selection.modelId,
-      chains: [{ id: 'chain-old', messages: fixtureMessages } as never] as never,
-    });
-    mocks.sessionManager._setModelHistory(fixtureMessages);
-    mocks.summarizeCompactableRange.mockResolvedValueOnce({ text: 'SUMMARY: scoped settle over eleven user turns' });
-    mocks.streamChat.mockImplementationOnce(textOnlyStream());
-
-    const send = vi.fn();
-    const chatSend = mocks.handlers.get(IPC_CHANNELS.CHAT_SEND)!;
-    await chatSend({ sender: { id: 970, send } }, { message: 'Follow up' });
-    await waitForDoneCount(send, 1);
-
-    // The send-time compaction fired and the summarizer covered the oldest
-    // user message — its only surviving representation is the summary head.
-    expect(mocks.summarizeCompactableRange).toHaveBeenCalledTimes(1);
-    const slice = mocks.summarizeCompactableRange.mock.calls[0]![0]!.messages as Array<{ id?: string }>;
-    expect(slice.some((m) => m.id === 'u-0')).toBe(true);
-
-    expect(mocks.streamChat).toHaveBeenCalledTimes(1);
-    const sentMessages = mocks.streamChat.mock.calls[0]![0]!.messages as Array<{
-      id?: string;
-      role?: string;
-      content?: string;
-      excludeFromModel?: boolean;
-      compacted?: unknown;
-    }>;
-    // Single representation: u-0 stays in the transcript but leaves the model
-    // view — no verbatim + summary double representation.
-    expect(sentMessages.find((m) => m.id === 'u-0')!.excludeFromModel).toBe(true);
-    // The ten newest user messages (plus the turn's own) survive verbatim.
-    for (let i = 1; i < 10; i += 1) {
-      expect(sentMessages.find((m) => m.id === `u-${i}`)!.excludeFromModel).not.toBe(true);
-    }
-    expect(sentMessages.some((m) => m.role === MessageRole.USER && m.content === 'Follow up' && !m.excludeFromModel)).toBe(true);
-    expect(sentMessages.some((m) => m.compacted && m.content === 'SUMMARY: scoped settle over eleven user turns')).toBe(true);
-
-    // Durable flags: u-0 and the in-range assistants flagged; exempt users not.
-    expect(mocks.sessionManager.applyCompaction).toHaveBeenCalledTimes(1);
-    const flaggedMessageIds = (mocks.sessionManager.applyCompaction.mock.calls[0]![1] as {
-      flaggedMessageIds: string[];
-    }).flaggedMessageIds;
-    expect(flaggedMessageIds).toContain('u-0');
-    expect(flaggedMessageIds).toContain('a-0');
-    for (let i = 1; i < 10; i += 1) {
-      expect(flaggedMessageIds).not.toContain(`u-${i}`);
-    }
-    expect(mocks.saveSession).not.toHaveBeenCalled();
   });
 });
